% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shinybrain.R
\name{add_shinybrain}
\alias{add_shinybrain}
\title{Initialize fresh shinybrain app}
\usage{
add_shinybrain(
  path,
  app_name = basename(path),
  overwrite = FALSE,
  open = FALSE,
  ...
)
}
\arguments{
\item{path}{Name of the folder to create the app in.
This will also be used as the app name.}

\item{app_name}{App name to use. By default, {shinybrain}
uses \code{basename(path)}. If \code{path == '.'} & \code{app_name} is
not explicitly set, then \code{basename(getwd())} will be used.}

\item{overwrite}{Boolean. Should the already existing app
be overwritten?}

\item{open}{Boolean. Open the created app project?}

\item{...}{additional arguments to be used later}
}
\value{
invisibly the path to new application
}
\description{
Initialize fresh shinybrain app
}
