% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shinylearning.R
\name{add_shinylearning}
\alias{add_shinylearning}
\title{Initialize fresh shiny learning app}
\usage{
add_shinylearning(app_name, path = getwd(), open = TRUE, ...)
}
\arguments{
\item{app_name}{string of the app name to use as the directory
to store application files}

\item{path}{parent directory to store app files. Default is the
current working directory}

\item{open}{Flag to open the project if using RStudio. Default
is TRUE.}

\item{...}{additional arguments to be used later}
}
\value{
invisibly the path to new application
}
\description{
Initialize fresh shiny learning app
}
