% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{app_page}
\alias{app_page}
\title{Compose single page of shinylearning app}
\usage{
app_page(nav_links, page_id = "page1", title = "myapp", bg = "#0062cc")
}
\arguments{
\item{nav_links}{list of navigation bar elements created with \code{gen_navitems}.}

\item{page_id}{page identifer.}

\item{title}{application title}

\item{bg}{color of the navigation bar}
}
\value{
list produced by \code{brochure::page}
}
\description{
Compose single page of shinylearning app
}
