% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{as_flag_list}
\alias{as_flag_list}
\title{Wrap a character vector into a list column value}
\usage{
as_flag_list(x = character())
}
\arguments{
\item{x}{Character vector or NULL}
}
\value{
List of length 1 containing x (suitable for list columns)
}
\description{
Wrap a character vector into a list column value
}
