% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{compute_usage_weights}
\alias{compute_usage_weights}
\title{Compute usage_count for helper_fn and state nodes}
\usage{
compute_usage_weights(nodes, edges)
}
\arguments{
\item{nodes}{sb_node_candidate tibble}

\item{edges}{sb_edge_candidate tibble}
}
\value{
Updated sb_node_candidate tibble
}
\description{
Compute usage_count for helper_fn and state nodes
}
