% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\docType{data}
\name{CONFIDENCE_LEVELS}
\alias{CONFIDENCE_LEVELS}
\title{Valid confidence levels}
\format{
An object of class \code{character} of length 3.
}
\usage{
CONFIDENCE_LEVELS
}
\description{
Valid confidence levels
}
\keyword{datasets}
