% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\docType{data}
\name{EDGE_TYPES}
\alias{EDGE_TYPES}
\title{Valid edge types}
\format{
An object of class \code{character} of length 10.
}
\usage{
EDGE_TYPES
}
\description{
Valid edge types
}
\keyword{datasets}
