% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_navbar.R
\name{gen_navitems}
\alias{gen_navitems}
\title{Generate navigation bar for shinylearning app}
\usage{
gen_navitems(path, active_tab = "home")
}
\arguments{
\item{path}{path to existing shinylearning app}

\item{active_tab}{which tab should be active when initializing the app.
Default is \code{home}.}
}
\value{
list of navigation bar elements created with \code{bslib}
}
\description{
Generate navigation bar for shinylearning app
}
