% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\docType{data}
\name{ISSUE_TYPES}
\alias{ISSUE_TYPES}
\title{Valid issue types}
\format{
An object of class \code{character} of length 9.
}
\usage{
ISSUE_TYPES
}
\description{
Valid issue types
}
\keyword{datasets}
