% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_sb_edge_candidate}
\alias{new_sb_edge_candidate}
\title{Create an empty sb_edge_candidate tibble}
\usage{
new_sb_edge_candidate()
}
\value{
tibble with 0 rows and correct column types
}
\description{
Create an empty sb_edge_candidate tibble
}
