% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_sb_symbol}
\alias{new_sb_symbol}
\title{Create an empty sb_symbol tibble}
\usage{
new_sb_symbol()
}
\value{
tibble with 0 rows and correct column types
}
\description{
Create an empty sb_symbol tibble
}
