% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{null_coalesce}
\alias{null_coalesce}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Primary value to return when it is not NULL and not length 0.}

\item{y}{Fallback value returned when \code{x} is NULL or empty.}
}
\value{
\code{x} when present, otherwise \code{y}.
}
\description{
Null coalescing operator
}
