% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_raw}
\alias{parse_raw}
\title{Parse a file and attach parse data}
\usage{
parse_raw(file_row)
}
\arguments{
\item{file_row}{One-row sb_file tibble (code must be populated)}
}
\value{
Named list: list(file = updated sb_file row, issues = sb_issue tibble)
}
\description{
Parse a file and attach parse data
}
