% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{print_brain_console}
\alias{print_brain_console}
\title{Print a formatted console report from an App Brain}
\usage{
print_brain_console(brain)
}
\arguments{
\item{brain}{App Brain from build_brain()}
}
\value{
Invisibly returns the brain.
}
\description{
Print a formatted console report from an App Brain
}
