% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\docType{data}
\name{REFERENCE_TYPES}
\alias{REFERENCE_TYPES}
\title{Valid reference types}
\format{
An object of class \code{character} of length 11.
}
\usage{
REFERENCE_TYPES
}
\description{
Valid reference types
}
\keyword{datasets}
