% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinybrain-package.R
\docType{package}
\name{shinybrain-package}
\alias{shinybrain}
\alias{shinybrain-package}
\title{shinybrain: Static analysis for Shiny applications}
\description{
Parses Shiny app source and builds a reactive-graph "brain" describing
contexts, symbols, references, nodes, and edges.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PrigasG/shinybrain}
  \item \url{https://prigasg.github.io/shinybrain/}
  \item Report bugs at \url{https://github.com/PrigasG/shinybrain/issues}
}

}
\author{
\strong{Maintainer}: ShinyBrain Authors \email{prigasgenthian48@gmail.com}

}
\keyword{internal}
