% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_shinybrain.R
\name{snapshot_shinybrain}
\alias{snapshot_shinybrain}
\title{Create new snapshot of the application}
\usage{
snapshot_shinybrain(
  path = getwd(),
  use_current_snapshot = TRUE,
  snapshot_name = NULL,
  style_code = TRUE,
  code_width = 30,
  ...
)
}
\arguments{
\item{path}{directory path to the application}

\item{use_current_snapshot}{Flag to use the current (or latest) snapshot of the
application as the source for the next snapshot. If set to \code{FALSE}, you must
supply a valid snapshot name for the \code{snapshot_name} parameter.}

\item{snapshot_name}{Optional character string for the name of the desired
snapshot to use as the basis for the new snapshot. If no value is supplied,
then by default it will use the current or latest snapshot available in the app.}

\item{style_code}{Flag to re-style the new snapshot code to meet tidy formatting
principles. Recommended to use the default value of \code{TRUE}.}

\item{code_width}{number of columns for each line in code files
produced by the snapshot. Ignored if \code{style_code} is disabled. Default
value is 30.}

\item{...}{additional arguments to be used later}
}
\value{
invisibly the path of the new snapshot
}
\description{
Create new snapshot of the application
}
