% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{validate_graph}
\alias{validate_graph}
\title{Validate graph integrity}
\usage{
validate_graph(nodes, edges, project_id)
}
\arguments{
\item{nodes}{sb_node_candidate tibble}

\item{edges}{sb_edge_candidate tibble}

\item{project_id}{Project ID}
}
\value{
sb_issue tibble (empty if no problems)
}
\description{
Checks for: orphan edges, duplicate node IDs, unrecognized node types
}
